Redirect to root_path on good ballot submission
This was a bit of a fun one to solve---redirecting to @ballot then
triggers a "You do not have the ability to access this resource."
error, which is what sticks.  So, users would submit a ballot,
immediately be directed to the #show action on that ballot, and then
be immediately redirected back to the root url with an error message.
(But only if they weren't admin.)

[GitHub] This commit resolves #11.